Ajustando hierarquia e permissões

+++ src/main/java/br/com/forum_hub/domain/resposta/RespostaService.java
@@ -4,7 +4,10 @@ import br.com.forum_hub.domain.topico.Status;
 import br.com.forum_hub.domain.topico.TopicoService;
 import br.com.forum_hub.domain.usuario.Usuario;
 import br.com.forum_hub.infra.exception.RegraDeNegocioException;
+import br.com.forum_hub.infra.security.HierarquiaService;
 import jakarta.transaction.Transactional;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.hierarchicalroles.RoleHierarchy;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,10 +16,12 @@ import java.util.List;
 public class RespostaService {
     private final RespostaRepository repository;
     private final TopicoService topicoService;
+    private final HierarquiaService hierarquiaService;
 
-    public RespostaService(RespostaRepository repository, TopicoService topicoService) {
+    public RespostaService(RespostaRepository repository, TopicoService topicoService, RoleHierarchy roleHierarchy, HierarquiaService hierarquiaService) {
         this.repository = repository;
         this.topicoService = topicoService;
+        this.hierarquiaService = hierarquiaService;
     }
 
     @Transactional
@@ -38,8 +43,11 @@ public class RespostaService {
     }
 
     @Transactional
-    public Resposta atualizar(DadosAtualizacaoResposta dados) {
+    public Resposta atualizar(DadosAtualizacaoResposta dados, Usuario logado) {
         var resposta = buscarPeloId(dados.id());
+        if(hierarquiaService.usuarioNaoTemPermissoes(logado, resposta.getAutor(), "ROLE_MODERADOR")) {
+            throw new AccessDeniedException("Você não pode apagar esse tópico!");
+        }
         return resposta.atualizarInformacoes(dados);
     }
 
@@ -48,21 +56,31 @@ public class RespostaService {
     }
 
     @Transactional
-    public Resposta marcarComoSolucao(Long id) {
+    public Resposta marcarComoSolucao(Long id, Usuario logado) {
         var resposta = buscarPeloId(id);
-
         var topico = resposta.getTopico();
-        if(topico.getStatus() == Status.RESOLVIDO)
-            throw new RegraDeNegocioException("O tópico já foi solucionado! Você não pode marcar mais de uma resposta como solução.");
+        if (resposta.ehSolucao()) {
+            throw new RegraDeNegocioException("Essa resposta já é solução do tópico: ID: " + topico.getId() + " Titulo: "+ topico.getTitulo());
+        }
 
+        if(hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_INSTRUTOR")) {
+            throw new AccessDeniedException("Você não pode marcar a resposta como solução");
+        }
+
+        if(topico.getStatus() == Status.RESOLVIDO) {
+            throw new RegraDeNegocioException("O tópico já foi solucionado! Você não pode marcar mais de uma resposta como solução.");
+        }
         topico.alterarStatus(Status.RESOLVIDO);
         return resposta.marcarComoSolucao();
     }
 
     @Transactional
-    public void excluir(Long id) {
+    public void excluir(Long id, Usuario logado) {
         var resposta = buscarPeloId(id);
         var topico = resposta.getTopico();
+
+        if(hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_MODERADOR"))
+            throw new AccessDeniedException("Você não pode apagar essa resposta!");
 
         repository.deleteById(id);
